separated react components
